🗺️ Atlas: Break circular dependency in stream and redaction
Extracted `RedactingSink` from `src/redaction.rs` to a new submodule
`src/stream/redacting.rs`. This breaks the cycle where `stream` exports
a struct implemented in `redaction.rs` which depends back on `stream`.
