+++ src/stream/mod.rs
@@ -15,6 +15,7 @@ use serde::Serialize;
 
 pub mod broadcast;
 pub mod event_log;
+pub mod redacting;
 pub mod sse;
 #[cfg(feature = "webhook")]
 pub mod sweep_webhook;
@@ -23,6 +24,8 @@ pub mod webhook;
 
 pub use broadcast::BroadcastSink;
 pub use event_log::EventLogSink;
+pub use redacting::RedactingSink;
+pub use redacting::redact_stream_event;
 pub use sse::SseServer;
 #[cfg(feature = "webhook")]
 pub use sweep_webhook::{
